Refactor entire codebase
Everything was refactored because the last time I touched this repo
I was essentially a newbie (read: did everything the wrong way) and
I hope that now I know a little bit more.

Code was restructured into 1 file (only 1 needed really), added a bunch
of checks that are super useful and added a few more utilities that
some users are going to enjoy.